requirements.txt
streamlit
pillow
opencv-python
numpy
requests
tensorflow==2.12
fpdf
